*Changes to satisfy Splint, Cppcheck, Clang's static analyzer and MinGW's pedantic level compile warnings. *Fixed a bug where if the last gem in a pro guitar note was toggled off when legacy view was in effect, the note would not be deleted.

+++ src/menu/note.c
@@ -2830,7 +2830,7 @@ static void eof_split_lyric(unsigned long lyric)
 	stringlen = strlen(eof_song->vocal_track[tracknum]->lyric[lyric]->text);
 
 	/* see how many pieces there are */
-	for(i = 0; i < stringlen; i++)
+	for(i = 0; i < (unsigned long) stringlen; i++)
 	{
 		lastc = c;
 		c = eof_song->vocal_track[tracknum]->lyric[lyric]->text[i];
